feat: add supabase client singleton
